fix: resolve CodeQL float-equality and loss-of-precision findings
- StrokeOutliner: use a small tolerance instead of exact-zero when
  guarding the outer/inner ring sign comparison (cs/equality-on-floats);
  the areas are computed values, not caller-supplied sentinels.
- PathBuilder.TangentArcTo: keep radius == 0f as an intentional
  exact-sentinel check (explicit caller request for 'no arc'), but use a
  tolerance for the computed len1/len2 near-coincident-point checks.
- StrokeOutlinerTests: mirror the same tolerance-based area comparison
  used in the production fix above.
- SurfaceTests.ComputeExpectedPremultiplied: cast the first operand to
  double before multiplying rather than casting the product afterward,
  resolving the cs/loss-of-precision finding about a possible integer
  multiplication overflow before widening.

Other flagged sites (SvgCodec.ParseDashArray, DashSplitter.IsSolidDash/
LocatePhaseForward/LocatePhaseBackward, GradientEvaluator.BuildPlan,
CornerRoundEffect.Apply, Path.RoundRectangle, SvgArcConverter.ToBeziers)
were reviewed and left unchanged: each compares a value against a
literal zero sentinel that is caller/spec-defined (an explicit radius
of zero, an explicit zero-length dash entry, or two directly
caller-supplied gradient centers/radii), not a computed value prone to
floating drift - converting these to tolerance comparisons would risk
changing documented, intentional behavior without fixing a real bug.
No suppression comments were added; findings judged intentional are
left as documented exact comparisons per review.

+++ test/DemaConsulting.CanvasNet.Tests/Canvas/SurfaceTests.cs
@@ -539,7 +539,11 @@ public class SurfaceTests
     /// </summary>
     private static byte ComputeExpectedPremultiplied(byte color, byte alpha)
     {
-        var value = Math.Round(color * alpha / 255.0, MidpointRounding.AwayFromZero);
+        // Cast the first operand to double before multiplying (rather than casting the product
+        // afterward) so the multiplication itself is performed in double precision - this is
+        // harmless for today's byte-range operands, but avoids the integer-multiplication-then-
+        // widen pattern that would silently truncate for wider integer types.
+        var value = Math.Round((double)color * alpha / 255.0, MidpointRounding.AwayFromZero);
         return (byte)Math.Clamp(value, 0, 255);
     }
